Move graphql stuff to separate file and use graphql handler
